Use u64 for the GroupWord on WebAssembly
Whether using 128-bit long SIMD vectors makes sense on WebAssembly is a
bit unclear with the bitmask operations on those not being able to be
lowered efficiently with NEON. However even the case without SIMD can be
improved as wasm32 is an architecture which has 32-bit pointers, but
native support for 64-bit integers. So the current general logic, that
uses the pointer size in most cases, needs to be special cased for
wasm32.