Provide default hasher types to `Vacant` and `Occupied` entries
For code which may select to use the `std` `HashMap` or the
`hashbrown` `HashMap` based on feature or cfg flags, being forced
to provide an explicit hash argument to `VacantEntry` or
`OccupiedEntry` structs requires jumping through some hoops.
Specifically, the `std` `hash_map::*Entry` structs don't take a
hasher type argument at all, making the downstream code change the
type entirely when switching from `std` to `hashbrown` or back.

For simplicity, its nice to have the argument be optional on the
`hashbrown` end, which would let the same code compile with either
map.