Add `HashSet::get_or_insert_with_mut`
This function avoids unnecessairy work when the value to be stored can
take ownership of the lookup key, such as when interning.
It makes a stronger guarantee than `HashSet::get_or_insert_with`, since
the reference given to `f` is guaranteed to be unique, and thus safe to
mutate.

The functionality of this function can already be achieved by (ab)using
`UnsafeCell`, but that relies on the implementation detail that the `&Q`
given to `f` is unique.