Introduce Swiss Cheese benchmark
* Changes:

- Introduce new swiss_cheese.rs benchmark.

* Motivation:

The current design of RawTable leaves tombstones behind to indicate
deleted elements, going from tightly packed elements to a much more
sparse design as more and more elements are deleted -- until more
elements are inserted again, at least.

This creates a "swiss cheese" situation, with a decreasing elements
density over time. This situation is expected to lead to a higher number
of cache misses as this density decreases.

The swiss_cheese.rs benchmark suite aims at measuring the impact of this
effect on performance.