Introduce Overflow & Displacement tracking.
* Changes:

- Introduce Overflow Trackers, with features to select the desired
  variant.
- Introduce Displacements, conditional on the Overflow Tracker variant
  tracking removals.
- Adjust insertion/removal of items in RawTable to properly track
  overflow and displacement.
- Adjust find in RawTable to short-circuit probe sequence when overflow
  tracking ensure there is no need to probe further.
- OF NOTE: enforce group alignment.

* Motivation:

Overflow tracking allows cutting a probing sequence short, which may be
beneficial.

The use of a multitude of variants makes it easier to test and benchmark
all variants, thus making it easier to pick the right one... or not pick
any.

The groups are now forcibly aligned because overflow tracking is
performed on a group basis, and does not work with "floating" groups.

* Design:

Overflow trackers and displacements are tacked at the end of the
allocation, and their access is minimized, so that their performance
impact is minimized.

In particular:

1. An element which does not overflow on insertion need not trigger a
   write to any overflow tracker, nor to its displacement.
2. Only if removals are tracked is the displacement read on removal.
3. Only if removals are tracked and the displacement is non-0 are
   overflow trackers written to on removal.

This follows the philosophy of "You Don't Pay For What You Don't Use",
and makes the impact as minimal as can be.