Remove the `raw` feature and make `RawTable` private
This will give more freedom for the internal implementation details of
hashbrown to evolve without the need for regular releases with breaking
changes.

All existing users of `RawTable` should migrate to the `HashTable` API
which is entirely safe while providing the same flexibility as
`RawTable`.

This also removes the following features which were only exposed under
`RawTable`:
- `RawTable::iter_hash`
- `RawIter::reflect_insert` and `RawIter::reflect_remove`
- `RawTable::clone_from_with_hasher`
- `RawTable::insert_no_grow` and `RawTable::try_insert_no_grow`
- `RawTable::allocation_info`
- `RawTable::try_with_capacity(_in)`
- `HashMap::raw_table(_mut)` and `HashSet::raw_table(_mut)`