Change `DefaultHashBuilder` to an opaque newtype
This hides the underlying hasher as an implementation detail, so it's no
longer a public dependency. A newtype `DefaultHasher` is also added here
to wrap the underlying `<_ as BuildHasher>::Hasher` type.

This is a breaking change now, requiring a semver bump, but it will
allow later changes to the hasher without breaking.